Don’t sort iterator output in test
Order between children of a trie is not defined (unless ‘enable_sorting’
is used of course) but order between parents and children is (namely,
iterator and listing methods use pre-order traversal) so the tests are
deterministic even if returned values are not sorted.

Even better, removing sorting tests that the pre-order traversal is
preserved.